+++ CHANGELOG.md
@@ -18,7 +18,8 @@ This file contains the list of changes made to pyjoulescope_ui.
 * Improved "developer" mode.
   * Renamed "debug" mode to "developer" mode.
   * Automatically close developer widgets on setting disable.
-* Added WindowStaysOnTopHint to HelpHtmlMessageBox  #245 
+* Added WindowStaysOnTopHint to HelpHtmlMessageBox  #245
+* Modified Export all data to work when streaming  #246
 
 
 ## 1.0.48